requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.7.2
django-taggit==0.23.0
django-widget-tweaks==1.4.3
idna==2.8
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1